+++ src/main/java/com/ncedu/fooddelivery/api/v1/entities/order/OrderNotHierarchical.java
@@ -62,15 +62,15 @@ public class OrderNotHierarchical {
 
     @NotNull
     @Column(name = "overall_cost")
-    private BigDecimal overallCost;
+    private Double overallCost;
 
     @NotNull
     @Column(name = "high_demand_coeff")
-    private BigDecimal highDemandCoeff;      // TODO: similarly "1.0"
+    private Double highDemandCoeff;      // TODO: similarly "1.0"
 
     @NotNull
     @Column(name = "discount")
-    private BigDecimal discount;     // TODO: similarly "0.0"
+    private Double discount;     // TODO: similarly "0.0"
 
     @Column(name = "promo_code_id")
     private Long promoCodeId;
@@ -83,7 +83,7 @@ public class OrderNotHierarchical {
 
     public OrderNotHierarchical(){}
 
-    public OrderNotHierarchical(Long id, Long clientId, String address, Geometry coordinates, Long warehouseId, Long courierId, OrderStatus status, Date dateStart, Date dateEnd, BigDecimal overallCost, BigDecimal highDemandCoeff, BigDecimal discount, Long promoCodeId, BigDecimal clientRating, BigDecimal deliveryRating) {
+    public OrderNotHierarchical(Long id, Long clientId, String address, Geometry coordinates, Long warehouseId, Long courierId, OrderStatus status, Date dateStart, Date dateEnd, Double overallCost, Double highDemandCoeff, Double discount, Long promoCodeId, BigDecimal clientRating, BigDecimal deliveryRating) {
         this.id = id;
         this.clientId = clientId;
         this.address = address;
